+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- new flag and config `interactive-page-size` to set how many entries should be shown on select prompts.
+
 ## [v0.43.0] - 2022-12-13
 
 ### Added